Binary Search Algorithm
Implemented binary search algorithm to search and check if the number is in the array of numbers. Note that the array should be in ascending order while passing the array to the binary search functions parameters else it might crash and won't be able to render the correct output. If you are passing an array in any other order make sure to sort the array and then pass the array to the binary search function.